refactor: rename constants file config to jm
