+++ src/hooks/CoinjoinRequirements.ts
@@ -1,6 +1,6 @@
 import * as fb from '../components/fb/utils'
 import { groupByJar, Utxos } from '../context/WalletContext'
-import { JM_TAKER_UTXO_AGE_DEFAULT } from '../constants/config'
+import { JM_TAKER_UTXO_AGE_DEFAULT } from '../constants/jm'
 
 export type CoinjoinRequirementOptions = {
   minNumberOfUtxos: number // min amount of utxos available
